+++ app/src/main/java/se/sugarest/jane/viaplaysections/data/datatype/package-info.java
@@ -1,0 +1,16 @@
+package se.sugarest.jane.viaplaysections.data.datatype;
+
+/**
+ * This package contains the data type used to parse the JSON from ViaplayAPI.
+ * <p>
+ * Complete Url example to request:
+ * https://content.viaplay.se/androidv2-se
+ * <p>
+ * Classes {@link se.sugarest.jane.viaplaysections.data.datatype.JSONResponse}
+ * {@link se.sugarest.jane.viaplaysections.data.datatype.Links}
+ * {@link se.sugarest.jane.viaplaysections.data.datatype.ViaplaySection}
+ * {@link se.sugarest.jane.viaplaysections.data.datatype.SingleJSONResponse} are automatically generated
+ * by http://www.jsonschema2pojo.org/
+ * <p>
+ * Created by jane on 17-11-16.
+ */
