--- app/src/main/java/se/sugarest/jane/viaplaysections/data/type/package-info.java
@@ -1,16 +1,0 @@
-package se.sugarest.jane.viaplaysections.data.type;
-
-/**
- * This package contains the data type used to parse the JSON from ViaplayAPI.
- * <p>
- * Complete Url example to request:
- * https://content.viaplay.se/androidv2-se
- * <p>
- * Classes {@link se.sugarest.jane.viaplaysections.data.type.JSONResponse}
- * {@link se.sugarest.jane.viaplaysections.data.type.Links}
- * {@link se.sugarest.jane.viaplaysections.data.type.ViaplaySection}
- * {@link se.sugarest.jane.viaplaysections.data.type.SingleJSONResponse} are automatically generated
- * by http://www.jsonschema2pojo.org/
- * <p>
- * Created by jane on 17-11-16.
- */
